Fix filename conflict in /tmp/
If two users tried to work on the same sketch, only the user that first
compiled the sketch would be able to work on it because they would own
the temp files.

This patch fixes that by including the md5sum of their user ID in the
sketch directory name.